Replace pircbot with kitteh (#12)
