+++ build.gradle
@@ -17,8 +17,8 @@ application {
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile group: 'org.mockito', name: 'mockito-core', version: '3.7.7'
-    implementation files('lib/pircbot.jar')
+    testImplementation group: 'junit', name: 'junit', version: '4.12'
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '3.7.7'
+    implementation "org.kitteh.irc:client-lib:8.0.0"
     implementation group: 'com.moandjiezana.toml', name: 'toml4j', version: '0.7.1'
 }
