+++ config.default.toml
@@ -1,5 +1,6 @@
 [irc]
 server = "irc.freenode.net"
+useSSL = false
 port = 6667
 nick = "PokerBot"
 channel = "#PokerBot"
