+++ src/main/java/me/ars/pokerbot/irc/Formatting.java
@@ -1,0 +1,40 @@
+package me.ars.pokerbot.irc;
+
+class Formatting {
+    /**
+     * Symbol for spades.
+     */
+    public static final String SPADES = "\u2660";
+    /**
+     * Symbol for hearts.
+     */
+    public static final String HEARTS = "\u2665";
+    /**
+     * Symbol for diamonds.
+     */
+    public static final String DIAMONDS = "\u2666";
+    /**
+     * Symbol for clubs.
+     */
+    public static final String CLUBS = "\u2663";
+    /**
+     * Clears formatting for the rest of the string.
+     */
+    public static final String CLEAR = "\u000f";
+    /**
+     * Makes the rest of the string bold.
+     */
+    public static final String BOLD = "\u0002";
+    /**
+     * Colors the rest of the string black.
+     */
+    public static final String COLOR_BLACK = "\u000301";
+    /**
+     * Colors the rest of the string red.
+     */
+    public static final String COLOR_RED = "\u000304";
+    /**
+     * Colors the rest of the string green.
+     */
+    public static final String COLOR_GREEN = "\u000303";
+}
