+++ src/main/java/me/ars/pokerbot/irc/Irc.java
@@ -1,0 +1,5 @@
+package me.ars.pokerbot.irc;
+
+public interface Irc {
+    void message(String channel, String message);
+}
